Support signature verification on models in Huggingface cache
Note: This requires model signining > v1.0.1, which is not available, yet.

Support signature verification on models donwloaded into the local
Huggingface cache. To be able to do signature verification on models there,
the model signing library requires to enable the following of symlinks,
for which the required API call will only be available in version >1.0.1.

Also, when downloading files from Huggingface, all files need to be
downloaded so that all files that are covered by the signature are
available when the signature is verified. Therefore, download all files if
necessary.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>