Enable signature verification on LoRA adapters
Enable signature verification on LoRA adapters loaded on startup as well
as those dynamically loaded during runtime. For this, extend the security
policy to allow for specifying LoRAs and their expected signers:

{
  "policy": {
    "signatures": {
      "signers": {
        "signer2": {
          "verification_method": "sigstore",
          "identity": "stefanb@us.ibm.com",
          "identity_provider": "https://sigstore.verify.ibm.com/oauth2"
        },
        "signer3": {
          "verification_method": "certificate",
          "certificate_chain": ["/home/stefanb/model-transparency/scripts/tests/keys/certificate/ca-cert.pem"]
        }
      },
      "models": {
        "/tmp/granite-3.3-8b-instruct/": {
          "signer": "signer2"
        }
      },
      "loras": {
         "/tmp/granite-3.3-8b-rag-agent-lib/query_rewrite_lora/" : {
             "signer": "signer3"
         },
         "regex:.*": {
             "signer": "signer2"
         }
      }
    }
  }
}

The above policy now enforces signature verification on the
query_rewrite_lora with the certificate method and all other LoRAs must
be signed with the sigstore method.

Make the SecurityPolicy class a 'dataclass' so it can be serialized. Add
a __post_init__ method it that will be called after de-serialization.

Extend the LoRARequest class to also carry the SecurityConfig that holds
the SecurityPolicy, which will be used for running signature verification
on the LoRA adapter.

Extend an existing test cases to test loading of unsigned and signed LoRA
adapters. For this, import a root and intermediate test CA and a signing
key for testing with the certificate/private key signing method.

With the above policy in /tmp/minimal-policy.json the LoRA adapter passed
on the command line as well as dynamically loaded LoRA adapters can have
their signatures enforced.

VLLM_PLUGINS=signature_enforcement \
VLLM_ALLOW_RUNTIME_LORA_UPDATING=True \
  vllm serve \
    --enforce-eager \
    --max-lora-rank 32 \
    /tmp/granite-3.3-8b-instruct/ \
    --enable-lora \
    --lora-modules \
       '{"name": "test", "path": "/tmp/granite-3.3-8b-rag-agent-lib/certainty_lora/"}' \
    --security-policy /tmp/minimal-policy.json

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>