Instrument lora_filesystem_resolver for signature verification
When a LoRA adapter is loaded using the completions API, the
file_system_resolver plugin may be used to located the LoRA adapter.
Instrument the code path leading to the LoRA adapter to pass through the
SecurityConfiguation and attach it to the LoRARequest for later signature
verification of the LoRA adapter. Add the SecurityConfigration parameter
to the LoRAResolver interface and adjust all other classes implementing
this interface to accept the new parameter.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>